pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "reachability-map"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "roboticstoolbox-python>=1.1.0",
    "spatialmath-python>=1.0.0",
    "numpy>=1.24.0,<2.0",
    "scipy>=1.11.0",
    "h5py>=3.9.0",
    "tqdm>=4.65.0",
]

[project.scripts]
reachability-map = "reachability_map.cli:main"

[tool.setuptools.packages.find]
where = ["."]
